// src/Enums/Categories.tsx
export enum Category {
    AnyCategory = "Any Category",
    GeneralKnowledge = "General Knowledge",
    Entertainment = "Entertainment",
    Science = "Science",
    Sports = "Sports",
    Geography = "Geography",
    HistoryAndPolitics = "History & Politics",
    HumanitiesAndArts = "Humanities & Arts"
} 